+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 23.4.0
 
+### New Modules
+  - na_ontap_autoupdate_config - REST only support for managing configurations for automatic updates, requires ONTAP 9.10.1 or later.
+
 ### Major Changes
   - na_ontap_snapmirror - AWS Lambda support added to the module.